feat: Migrate to a Core Validation Module
This has decoupled the SMTP session from the actual validation.
Effectively, this serves the purpose of having a saner distribution of
concerns,
but does also allow one to change without effecting the other.
Additionally, this means that any changes needed for the validation
to occur, would necessitate exposing it to user modules as well.